Express server & mongoose initialized, POST implemented for /api/expenses
