+++ README.md
@@ -1,1 +1,1 @@
-The Backend Server for expense app based on flask
+RESR Server for ExpenseApp
