+++ package.json
@@ -1,26 +1,28 @@
 {
   "name": "expense-backend",
   "version": "0.0.1",
-  "description": "Backend for expense app",
+  "description": "REST server for ExpenseApp",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/Praseetha-KR/expense-backend.git"
+    "url": "git+https://github.com/Praseetha-KR/expense-backend.git"
   },
+  "keywords": [
+    "rest",
+    "expense"
+  ],
   "author": "Praseetha-KR",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/Praseetha-KR/expense-backend/issues"
   },
-  "homepage": "https://github.com/Praseetha-KR/expense-backend",
+  "homepage": "https://github.com/Praseetha-KR/expense-backend#readme",
   "dependencies": {
-    "body-parser": "^1.9.3",
-    "cors": "^2.5.2",
-    "express": "^4.10.4",
-    "mongoose": "^3.8.20"
+    "body-parser": "^1.14.1",
+    "express": "^4.13.3",
+    "mongoose": "^4.2.2"
   }
 }
